Support RV64 builds to use with spike.
OpenOCD does not currently support debugging 64-bit FreeRTOS, but now
that I have a target to test hopefully that will be remedied shortly.

+++ FreeRTOS/Demo/RISC-V-spike-htif_GCC/main_blinky.c
@@ -119,7 +119,7 @@ int main_blinky( void )
 	vSendString( "Hello FreeRTOS!" );
 
 	/* Create the queue. */
-	xQueue = xQueueCreate( mainQUEUE_LENGTH, sizeof( uint32_t ) );
+	xQueue = xQueueCreate( mainQUEUE_LENGTH, sizeof( unsigned long ) );
 
 	if( xQueue != NULL )
 	{
